webcrawler_agent/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
PyYAML==6.0